+++ packages/desktop/src/renderer/components/ui/OpenlinkModal/OpenlinkModal.test.tsx
@@ -24,8 +24,9 @@ describe('OpenlinkModal', () => {
           aria-hidden="true"
         />
         <div
-          class="Modalroot MuiModal-root css-1voaj9u-MuiModal-root"
+          class="MuiModal-root css-1vjugmr-MuiModal-root"
           role="presentation"
+          zindex="1300"
         >
           <div
             aria-hidden="true"