+++ packages/desktop/src/renderer/components/widgets/update/UpdateModal.test.tsx
@@ -14,8 +14,9 @@ describe('UpdateModal', () => {
           aria-hidden="true"
         />
         <div
-          class="Modalroot MuiModal-root css-1voaj9u-MuiModal-root"
+          class="MuiModal-root css-1vjugmr-MuiModal-root"
           role="presentation"
+          zindex="1300"
         >
           <div
             aria-hidden="true"