+++ packages/e2e-tests/src/tests/oneClient.test.ts
@@ -124,12 +124,6 @@ describe('One Client', () => {
         })
       }
 
-      it('User waits for the modal StartingLoadingPanel to disappear', async () => {
-        const loadingPanel = new StartingLoadingPanel(app.driver)
-        const isLoadingPanel = await loadingPanel.element.isDisplayed()
-        expect(isLoadingPanel).toBeTruthy()
-      })
-
       it('User sees "general channel" page', async () => {
         const generalChannel = new Channel(app.driver, 'general')
         const isGeneralChannel = await generalChannel.element.isDisplayed()