+++ packages/eslint-config-custom/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.3.0](/compare/@quiet/eslint-config@1.3.0-alpha.0...@quiet/eslint-config@1.3.0) (2023-08-28)
+
+**Note:** Version bump only for package @quiet/eslint-config
+
+
+
+
+
+# [1.3.0-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/eslint-config@1.2.0...@quiet/eslint-config@1.3.0-alpha.0) (2023-08-25)
+
+**Note:** Version bump only for package @quiet/eslint-config
+
+
+
+
+
+# [1.2.0](https://github.com/TryQuiet/quiet/compare/@quiet/eslint-config@1.1.0...@quiet/eslint-config@1.2.0) (2023-08-17)
+
+**Note:** Version bump only for package @quiet/eslint-config
+
+
+
+
+
 # [1.1.0](https://github.com/ZbayApp/monorepo/compare/@quiet/eslint-config@1.0.1-alpha.0...@quiet/eslint-config@1.1.0) (2023-07-28)
 
 **Note:** Version bump only for package @quiet/eslint-config